% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeCindex}
\alias{computeCindex}
\title{Calculate C-index (consistency index)}
\usage{
computeCindex(beta_6, data_1)
}
\arguments{
\item{beta_6}{the estimator that has been obtained from AFT model}

\item{data_1}{data in matrix form}
}
\value{
Returns a number between 0-1
}
\description{
Calculate C-index (consistency index)
}
\examples{
\dontrun{
    result <- computeCindex(as.numeric(beta_0), as.matrix(data_r))
}
}
