% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcntion.R
\name{f_gener_data}
\alias{f_gener_data}
\title{Generate heteroscedastic survival data}
\usage{
f_gener_data(N, p, type, heteroskedasticity = TRUE)
}
\arguments{
\item{N}{the number of samples}

\item{p}{dimension of samples}

\item{type}{The type of error distribution selected,"normal","extreme"}

\item{heteroskedasticity}{Whether the generated data has variance heterogeneity, defaults to TRUE}
}
\value{
dataframe
}
\description{
Generate heteroscedastic survival data
}
\examples{
\dontrun{
        set.seed(100)
        N<-500
        p<-2
        data_1 <- f_gener_data(N,p,"extreme",heteroskedasticity=TRUE)
}
}
