% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcntion.R
\name{f_update_basiscoef}
\alias{f_update_basiscoef}
\title{Obtain non-parametric estimates of tau quantiles}
\usage{
f_update_basiscoef(tau, data, beta, tau_pre, beta_pre)
}
\arguments{
\item{tau}{The quantiles you want to estimate}

\item{data}{survival data}

\item{beta}{the estimator that has been obtained from AFT model}

\item{tau_pre}{Previous quantiles, vector}

\item{beta_pre}{The previous quantile estimator,Matrix,one column corresponds to a quantile}
}
\value{
vector
}
\description{
This function will return the coefficients under the spline basis function, and thus the user can obtain the tau quantile non-parametric estimator of the survival data by himself. And when using this function, you need to know those quantiles that are smaller than tau
}
\examples{
\dontrun{
     set.seed(100)
     N<-500
     p<-2
     data_1 <- f_gener_simple(N,p,"extreme")
     library(aftgee)
     library(survival)
     fit1 <- aftgee::aftgee(Surv(Y, status) ~ ., data = data_1)
     beta_0 <-matrix(fit1$coefficients[,1])
     tau_0<-0.04
     tau_list <- matrix(sort(seq(0,1-tau_0,tau_0), decreasing = TRUE),nrow = 1)
     basis_number<-f_get_basisnumber(data_main,beta_0)
     tau_pre_out<-matrix(c(0),nrow=1)
     beta_pre_out<-matrix(c(rep(-10,basis_number)),nrow=basis_number)
     #initial
     for (j in 1:(length(tau_list)-1)) 
     {
     beta_up<-f_update_basiscoef(tau=tau_list[1,length(tau_list)-j],
     data=data_1,beta=beta_0,
     tau_pre=tau_pre_out,beta_pre=beta_pre_out)
     beta_pre_out <- cbind(beta_up,beta_pre_out)
     tau_pre_out <- matrix(tau_list[1,(length(tau_list)-j):length(tau_list)],nrow=1)
     }
}
}
